fix: task assignment API compatibility, add All Employees option, add sprints description column, and migration DATE_PART fixes
- Map assigned_to to assignees array in task creation request
- Add 'All Employees' option to task assignee dropdown
- Add missing description column to sprints table via migration
- Update Sprint TypeScript interface with description field
- Fix DATE_PART numeric casts in attendance/break migrations

+++ supabase/migrations/20260728_attendance_breaks.sql
@@ -61,7 +61,7 @@ CREATE OR REPLACE FUNCTION calculate_break_duration()
 RETURNS TRIGGER AS $$
 BEGIN
   IF NEW.start_time IS NOT NULL AND NEW.end_time IS NOT NULL THEN
-    NEW.duration := ROUND(DATE_PART('EPOCH', NEW.end_time - NEW.start_time) / 60.0, 2);
+    NEW.duration := ROUND((DATE_PART('EPOCH', NEW.end_time - NEW.start_time) / 60.0)::numeric, 2);
   ELSE
     NEW.duration := NULL;
   END IF;
