+++ supabase/migrations/20260728_half_day_auto.sql
@@ -21,7 +21,7 @@ DECLARE
   net_hours NUMERIC(4,2);
 BEGIN
   IF NEW.check_in_time IS NOT NULL AND NEW.check_out_time IS NOT NULL THEN
-    raw_hours := ROUND(DATE_PART('EPOCH', NEW.check_out_time - NEW.check_in_time) / 3600.0, 2);
+    raw_hours := ROUND((DATE_PART('EPOCH', NEW.check_out_time - NEW.check_in_time) / 3600.0)::numeric, 2);
 
     -- Look up shift start/end for this employee on this date
     SELECT s.start_time::time, s.end_time::time
@@ -76,7 +76,7 @@ BEGIN
       NEW.early_departure_minutes := 0;
     ELSIF shift_duration_hours > 0 AND NEW.check_out_time::TIME < shift_end THEN
       NEW.early_departure := TRUE;
-      NEW.early_departure_minutes := GREATEST(ROUND(DATE_PART('EPOCH', shift_end - NEW.check_out_time::TIME) / 60.0), 0);
+      NEW.early_departure_minutes := GREATEST(ROUND((DATE_PART('EPOCH', shift_end - NEW.check_out_time::TIME) / 60.0)::numeric), 0);
     END IF;
   ELSE
     NEW.working_hours := 0;
